Move RapidAPI stuff under ext.rapidapi
As it is not exactly a provider.
